# openapi.yaml
openapi: 3.0.3
info:
  title: Pafodex Backend API
  version: 1.0.0
  description: |
    Documentation Swagger de l'API Pafodex.
    Les endpoints exposés sont détaillés avec les payloads, paramètres, réponses et pagination.
servers:
  - url: /
    description: Serveur courant
  - url: https://pafodex-backend.onrender.com
    description: Serveur Render
paths:
...
  /api/cards:
    get:
      tags:
        - Cards
      summary: Lister les cartes
      parameters:
        - name: page
          in: query
          description: Numéro de page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Nombre de cartes par page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
      responses:
        '200':
          description: Liste paginée des cartes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardsListResponse'
    post:
      tags:
        - Cards
      summary: Créer une carte
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        '201':
          description: Carte créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '400':
          description: Requête invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: GameType introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/card/{id}:
    get:
      tags:
        - Cards
      summary: Récupérer une carte par ID
      parameters:
        - name: id
          in: path
          description: Identifiant de la carte
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Détails de la carte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '404':
          description: Carte introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Cards
      summary: Mettre à jour une carte
      parameters:
        - name: id
          in: path
          description: Identifiant de la carte
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCardRequest'
      responses:
        '200':
          description: Carte mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '400':
          description: Requête invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Carte ou GameType introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Cards
      summary: Supprimer une carte
      parameters:
        - name: id
          in: path
          description: Identifiant de la carte
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Suppression réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Carte introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/me/sets/{setId}/card:
    post:
      tags:
        - Sets
      summary: Ajouter une carte à un set
      description: |
        Le token peut être fourni dans l'en-tête Authorization: Bearer <token>
        ou en query paramètre apiToken.
      parameters:
        - $ref: '#/components/parameters/ApiTokenQuery'
        - name: setId
          in: path
          description: Identifiant du set
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCardToSetRequest'
      responses:
        '200':
          description: Carte ajoutée au set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '400':
          description: Requête invalide ou type incompatible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token manquant ou invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Accès refusé au set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Set ou carte introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/me/sets/{setId}/card/{cardId}:
    delete:
      tags:
        - Sets
      summary: Retirer une carte d'un set
      description: |
        Le token peut être fourni dans l'en-tête Authorization: Bearer <token>
        ou en query paramètre apiToken.
      parameters:
        - $ref: '#/components/parameters/ApiTokenQuery'
        - name: setId
          in: path
          description: Identifiant du set
          required: true
          schema:
            type: integer
        - name: cardId
          in: path
          description: Identifiant de la carte
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Carte retirée du set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Token manquant ou invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Accès refusé au set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Set ou carte introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Bearer
  parameters:
    ApiTokenQuery:
      name: apiToken
      in: query
      description: Token d'API alternatif lorsque Authorization n'est pas fourni
      required: false
      schema:
        type: string
  schemas:
...
    CreateCardRequest:
      type: object
      required:
        - name
        - extension
        - number
        - gameTypeId
      properties:
        name:
          type: string
          example: Pikachu
        extension:
          type: string
          example: Base
        number:
          type: string
          example: "001"
        image:
          type: string
          format: uri
          nullable: true
          example: https://example.com/image.png
        gameTypeId:
          type: integer
          example: 1
    UpdateCardRequest:
      type: object
      properties:
        name:
          type: string
          example: Pikachu
        extension:
          type: string
          example: Base
        number:
          type: string
          example: "001"
        image:
          type: string
          format: uri
          nullable: true
          example: https://example.com/image.png
        gameTypeId:
          type: integer
          example: 1
    CardResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Pikachu
        extension:
          type: string
          example: Base
        number:
          type: string
          example: "001"
        image:
          type: string
          format: uri
          nullable: true
          example: https://example.com/image.png
        gameType:
          $ref: '#/components/schemas/GameTypeSummary'
    CardsListResponse:
      type: object
      properties:
        cards:
          type: array
          items:
            $ref: '#/components/schemas/CardResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
    GameTypeSummary:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nom:
          type: string
          example: Pokemon
...
    AddCardToSetRequest:
      type: object
      required:
        - cardId
      properties:
        cardId:
          type: integer
          example: 10
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: Operation completed successfully
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        perPage:
          type: integer
          example: 25
        total:
          type: integer
          example: 100
        pages:
          type: integer
          example: 4
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Missing fields